Variadic functions

+++ 0x10-variadic_functions/3-print_all.c
@@ -30,7 +30,7 @@ void format_int(char *separator, va_list ap)
 
 void format_float(char *separator, va_list ap)
 {
-	printf("%s%f", separator, va_arg(ap, int));
+	printf("%s%f", separator, va_arg(ap, double));
 }
 
 /**
